seperate class from tkobject
