+++ ToolKit/ToolKit.vcxproj.filters
@@ -25,6 +25,9 @@
     <Filter Include="Source\Render\PostProcessPass">
       <UniqueIdentifier>{d2993129-0199-49c6-91fd-e82cc756b97f}</UniqueIdentifier>
     </Filter>
+    <Filter Include="Source\Reflection">
+      <UniqueIdentifier>{be940542-2f55-45a4-9c61-3f2c593d7c45}</UniqueIdentifier>
+    </Filter>
   </ItemGroup>
   <ItemGroup>
     <ClCompile Include="Drawable.cpp">
@@ -37,9 +40,6 @@
       <Filter>Source</Filter>
     </ClCompile>
     <ClCompile Include="Node.cpp">
-      <Filter>Source</Filter>
-    </ClCompile>
-    <ClCompile Include="ParameterBlock.cpp">
       <Filter>Source</Filter>
     </ClCompile>
     <ClCompile Include="PluginManager.cpp">
@@ -225,16 +225,19 @@
     <ClCompile Include="GlErrorReporter.cpp">
       <Filter>Source\Render</Filter>
     </ClCompile>
-    <ClCompile Include="TKObject.cpp">
-      <Filter>Source</Filter>
-    </ClCompile>
     <ClCompile Include="EngineSettings.cpp">
       <Filter>Source</Filter>
     </ClCompile>
     <ClCompile Include="ForwardPreProcessPass.cpp" />
     <ClCompile Include="AdditiveLightingPass.cpp" />
+    <ClCompile Include="ParameterBlock.cpp">
+      <Filter>Source\Reflection</Filter>
+    </ClCompile>
     <ClCompile Include="ObjectFactory.cpp">
-      <Filter>Source</Filter>
+      <Filter>Source\Reflection</Filter>
+    </ClCompile>
+    <ClCompile Include="TKObject.cpp">
+      <Filter>Source\Reflection</Filter>
     </ClCompile>
   </ItemGroup>
   <ItemGroup>
@@ -257,9 +260,6 @@
       <Filter>Source</Filter>
     </ClInclude>
     <ClInclude Include="Node.h">
-      <Filter>Source</Filter>
-    </ClInclude>
-    <ClInclude Include="ParameterBlock.h">
       <Filter>Source</Filter>
     </ClInclude>
     <ClInclude Include="Plugin.h">
@@ -457,16 +457,22 @@
     <ClInclude Include="GlErrorReporter.h">
       <Filter>Source\Render</Filter>
     </ClInclude>
-    <ClInclude Include="TKObject.h">
-      <Filter>Source</Filter>
-    </ClInclude>
     <ClInclude Include="EngineSettings.h">
       <Filter>Source</Filter>
     </ClInclude>
     <ClInclude Include="ForwardPreProcessPass.h" />
     <ClInclude Include="AdditiveLightingPass.h" />
+    <ClInclude Include="ParameterBlock.h">
+      <Filter>Source\Reflection</Filter>
+    </ClInclude>
     <ClInclude Include="ObjectFactory.h">
-      <Filter>Source</Filter>
+      <Filter>Source\Reflection</Filter>
+    </ClInclude>
+    <ClInclude Include="TKObject.h">
+      <Filter>Source\Reflection</Filter>
+    </ClInclude>
+    <ClInclude Include="Class.h">
+      <Filter>Source\Reflection</Filter>
     </ClInclude>
   </ItemGroup>
 </Project>
